pyproject.toml
[project]
name = "cogitarelink-solid"
version = "0.1.0"
description = "Solid Pod as a knowledge fabric node: bridging document and graph access via navigational meta-structure"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [
    { name = "Charles F. Vardeman II", email = "cvardema@nd.edu" },
]
dependencies = [
    "httpx>=0.28",
    "rdflib>=7.0",
    "pyshacl>=0.31",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0",
    "pytest-httpx>=0.35",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests/pytest"]
markers = ["integration: tests requiring a running CSS pod"]
